requirements.txt
ccxt>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.12.0
apscheduler>=3.10.0
requests>=2.31.0
python-telegram-bot>=20.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
aiofiles>=23.2.0
psycopg2-binary>=2.9.0
pydantic-settings>=2.0.0
passlib>=1.7.4
bcrypt==4.0.1
pyotp>=2.9.0
qrcode>=7.4.0
pillow>=10.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
fastapi-users>=12.1.0
fastapi-users-db-sqlalchemy>=6.0.0
pyyaml>=6.0.0
dnspython>=2.0.0
web3>=6.0.0
solana>=0.30.0
plotly>=5.0.0
psutil>=5.9.0
